refactor(web): move avatar-api.ts from utils/ to assistant/ (has I/O side effects)
avatar-api.ts makes HTTP requests via the HeyAPI client — it doesn't
belong in utils/ (pure functions only). The avatar API endpoints are all
/v1/assistants/{id}/avatar/... so the natural home is assistant/,
alongside assistant/api.ts and assistant/identity.ts.

Addresses Devin Review finding on PR #32185.
