+++ apps/web/src/assistant/avatar-api.ts
